removing suggest_nearby, replacing suggest with suggest_multiple, commenting out a couple of tests for right now.
